Move converters to json folder
