+++ src/api/Program.cs
@@ -2,6 +2,7 @@
 using Amazon.S3;
 using BDMS;
 using BDMS.Authentication;
+using BDMS.Json;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc;
