+++ tests/extern-sync/TestSyncContextExtensions.cs
@@ -1,4 +1,5 @@
-﻿using BDMS.Models;
+﻿using BDMS.Json;
+using BDMS.Models;
 using DotNet.Testcontainers.Builders;
 using Microsoft.EntityFrameworkCore;
 using NetTopologySuite.IO.Converters;
